pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["m-noman-ch <noman@aitomation.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework-simplejwt = "^5.5.0"
djangorestframework = "^3.16.0"
pillow = "^11.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
